Initial migration to net6
